refactor(auth): move getCompleteBalances method
