+++ lib/authenticated.js
@@ -2,18 +2,6 @@ const PublicClient = require("./public.js");
 
 class AuthenticatedClient extends PublicClient {
   /**
-   * @param {Object} [options]
-   * @param {string} [options.account] - Set the `account` parameter to "all" to include your margin and lending accounts.
-   * @example
-   * const balances = authClient.getCompleteBalances({ account: 'all' });
-   * @see {@link https://docs.poloniex.com/?shell#returncompletebalances|returnCompleteBalances}
-   * @description Get all of your balances, including available balance, balance on orders, and the estimated BTC value of your balance.
-   */
-  getCompleteBalances({ account } = {}) {
-    return this.post({ command: "returnCompleteBalances", account });
-  }
-
-  /**
    * @example
    * const addresses = authClient.getDepositAddresses();
    * @see {@link https://docs.poloniex.com/?shell#returndepositaddresses|returnDepositAddresses}
