+++ tests/authenticated.spec.js
@@ -10,36 +10,6 @@ const secret = "poloniex-api-secret";
 const authClient = new Poloniex.AuthenticatedClient({ key, secret });
 
 suite("AuthenticatedClient", () => {
-  test(".getCompleteBalances()", done => {
-    const balances = {
-      BTC: {
-        available: "0.00000000",
-        onOrders: "0.00000000",
-        btcValue: "0.00000000"
-      },
-      USDT: {
-        available: "0.00000000",
-        onOrders: "0.00000000",
-        btcValue: "0.00000000"
-      }
-    };
-    const nonce = 1560742707669;
-    authClient.nonce = () => nonce;
-
-    nock(EXCHANGE_API_URL)
-      .post("/tradingApi", { command: "returnCompleteBalances", nonce })
-      .times(1)
-      .reply(200, balances);
-
-    authClient
-      .getCompleteBalances()
-      .then(data => {
-        assert.deepStrictEqual(data, balances);
-        done();
-      })
-      .catch(error => assert.fail(error));
-  });
-
   test(".getCompleteBalances() (with parameters)", done => {
     const options = { account: "all" };
     const balances = {
